Fix manager routes and docs
Fix the routes of the manager modules and document some js functions

+++ Frontend/views/js/modules/manager/enrollments.js
@@ -40,8 +40,14 @@ $(document).ready(function () {
   });
 });
 
+/**
+ * Function to generate a random number between a set minimum and maximum number, including them
+ * @param {number} min - Minimum value
+ * @param {number} max - Maximum value
+ * @returns A random number between the 'min' and 'max' values, rounded to 0 decimals and parsed as an integer
+ */
 function generateRandomNumber(min, max) {
-  return parseFloat((Math.random() * (max - min) + min).toFixed(0));
+  return parseInt((Math.random() * (max - min) + min).toFixed(0));
 }
